docs: clarify content/ vs docs/ role split in contribution guides
